+++ watch/stac/__init__.py
@@ -1,0 +1,3 @@
+"""
+Collections of tooling related to STAC to support the CLI
+"""