+++ CHANGELOG.md
@@ -4,6 +4,37 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
+## Version 0.3.4 - Unreleased
+
+### Added
+* New "space_scale" parameter to Dataset such that a specific GSD can be given on the fly
+* New samecolor heuristic to remove invalid regions in data sampling.
+* Can now force nodata values in prep ta2 and align script.
+
+### Changed
+* The nir08 sensor for L8 is now aliased to nir for consistency.
+* Add support for lightning 1.7
+
+
+## Version 0.3.3 - Unreleased
+
+### Changed
+* Updates to handle new "Drop4" datasets
+* Fixes to sensorchan integration with the model and dataset (training with new settings is now verified)
+* Fixes to handling of nan data
+* Fixes to lightning modules
+* Fixed issue where learning_rate was not respected by some optimizers (RADam, but AdamW seemed ok)
+* Fixed issue where weight_decay was not respected by AdamW.
+* One-off scripts for fixing models.
+* Better handling of gdal warp / translate / mosaic failures.
+* Moved monkey patches to their own module.
+* Other minor fixes.
+
+## Version 0.3.2 - Unreleased
+
+Many undocumented changes
+
+
 ## Version 0.0.1 - Unreleased